Update 3-26-2016
No longer utilizing TimerOne library. Eliminating it in favor for an
external oscillator (MXO45HS-3C-1M0000). It has the capability to be
enabled/disabled by the MCU (not currently utilized) for power savings
in low power situations.

Also utilizing AltSoftSerial as it no longer interferes with TimerOne.
AltSoftSerial has the capability to send and receive data
simultaneously, but with some caveats not presently realized.
SoftwareSerial works fine as is and can be continued to use if desired.